requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
Django==4.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.0
executing==2.0.1
Faker==24.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
ipython==8.23.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
openai==1.23.6
parso==0.8.4
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
toposort==1.10
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
